Added MegaLoc and removed EigenPlaces
