+++ hloc/extract_features.py
@@ -141,9 +141,9 @@ confs = {
         "model": {"name": "openibl"},
         "preprocessing": {"resize_max": 1024},
     },
-    "eigenplaces": {
-        "output": "global-feats-eigenplaces",
-        "model": {"name": "eigenplaces"},
+    "megaloc": {
+        "output": "global-feats-megaloc",
+        "model": {"name": "megaloc"},
         "preprocessing": {"resize_max": 1024},
     },
 }
